refactor(arch): extract zone-addressed APIs into ArchSpecGeometry
The #419/#420 zone-addressed helpers (get_zone_grid, get_all_sites,
get_available_buses, get_grid_endpoints) plus BusDescriptor are
Python-side computation that produces bloqade.geometry grids and
Python-friendly bus descriptors. Move them into a new
ArchSpecGeometry helper class in arch_geometry.py so the core
ArchSpec wrapper stays close to a thin shell over the Rust spec
in preparation for eventually replacing the Python wrapper with
the Rust spec directly.

The new class accepts either a Python ArchSpec or a Rust
_RustArchSpec, so downstream consumers can build it from either
layer.
